Renamed EventsBuilder to EventBuilder
