+++ src/platforms/alexa/apis/ProactiveEvents.ts
@@ -20,17 +20,17 @@
  * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-import * as _ from "lodash";
 import * as rp from "request-promise";
 
 import { AuthenticationBase } from "./AuthenticationBase";
+import { EventBuilder } from "./proactiveEventBuilders/EventBuilder";
 
 export class ProactiveEvents extends AuthenticationBase {
   /**
    * Creates proactive event
    * https://developer.amazon.com/docs/smapi/proactive-events-api.html
    */
-  public async createEvent(endpoint: string, body: EventsBuilder, isDevelopment?: boolean): Promise<any> {
+  public async createEvent(endpoint: string, body: EventBuilder, isDevelopment?: boolean): Promise<any> {
     const tokenResponse = await this.getAuthenticationToken("alexa::proactive_events");
     let uri = `${endpoint}/v1/proactiveEvents`;
 
@@ -50,104 +50,5 @@ export class ProactiveEvents extends AuthenticationBase {
     };
 
     return Promise.resolve(rp(options));
-  }
-}
-
-/**
- * Events Builder class reference
- */
-export class EventsBuilder {
-  private expiryTime: string = "";
-  private timestamp: string = "";
-  private localizedAttributes: any[] = [];
-  private name: string|undefined;
-  private payload: any = {};
-  private referenceId: string = "";
-  private relevantAudience: any = {};
-
-  constructor(name?: string) {
-    this.name = name;
-  }
-
-  public addContent(locale: string, localizedKey: string, localizedValue: string): EventsBuilder {
-    this.localizedAttributes = this.localizedAttributes || [];
-
-    let item: any = _.find(this.localizedAttributes, (x) => x.locale === locale);
-
-    if (!item) {
-      item = { locale };
-      item[localizedKey] = localizedValue;
-      this.localizedAttributes.push(item);
-    } else {
-      item[localizedKey] = localizedValue;
-    }
-
-    return this;
-  }
-
-  /*
-   * expiryTime must be in GMT
-   */
-  public setExpiryTime(expiryTime: string): EventsBuilder {
-    this.expiryTime = expiryTime;
-
-    return this;
-  }
-
-  public setPayload(payload: any): EventsBuilder {
-    this.payload = payload;
-
-    return this;
-  }
-
-  public setReferenceId(referenceId: string): EventsBuilder {
-    this.referenceId = referenceId;
-
-    return this;
-  }
-
-  public setMulticast(): EventsBuilder {
-    this.relevantAudience = {
-      payload: {},
-      type: "Multicast",
-    };
-
-    return this;
-  }
-
-  /*
-   * timestamp must be in GMT
-   */
-  public setTimestamp(timestamp: string): EventsBuilder {
-    this.timestamp = timestamp;
-
-    return this;
-  }
-
-  public setUnicast(userId: string): EventsBuilder {
-    this.relevantAudience = {
-      payload: { user: userId },
-      type: "Unicast",
-    };
-
-    return this;
-  }
-
-  public getPayload() {
-    return this.payload;
-  }
-
-  public build(): any {
-    return {
-      event: {
-        name: this.name,
-        payload: this.getPayload(),
-      },
-      expiryTime: this.expiryTime,
-      localizedAttributes: this.localizedAttributes,
-      referenceId: this.referenceId,
-      relevantAudience: this.relevantAudience,
-      timestamp: this.timestamp,
-    };
   }
 }
